完善简历科技代码

+++ src/main/java/org/lsqt/rst/controller/WorkAddressController.java
@@ -10,7 +10,9 @@ import org.lsqt.components.context.annotation.Inject;
 import org.lsqt.components.context.annotation.mvc.RequestMapping;
 
 import org.lsqt.components.db.Db;
+import org.lsqt.components.db.IdGenerator;
 import org.lsqt.components.db.Page;
+import org.lsqt.components.db.orm.ORMappingIdGenerator;
 import org.lsqt.components.util.lang.StringUtil;
 import org.lsqt.sys.model.Dictionary;
 import org.lsqt.sys.service.DictionaryService;
@@ -28,7 +30,7 @@ import org.lsqt.rst.service.WorkAddressService;
  */
 @Controller(mapping={"/rst/work_address"})
 public class WorkAddressController {
-	
+	private IdGenerator idgen = new ORMappingIdGenerator();
 	@Inject private WorkAddressService workAddressService; 
 	
 	@Inject private Db db;
@@ -41,7 +43,12 @@ public class WorkAddressController {
 	
 	@RequestMapping(mapping = { "/page", "/m/page" })
 	public Page<WorkAddress> queryForPage(WorkAddressQuery query) throws IOException {
-		return workAddressService.queryForPage(query); //  
+		return workAddressService.queryForPage(query); 
+	}
+	
+	@RequestMapping(mapping = { "/list", "/m/list" })
+	public List<WorkAddress> queryForLits(WorkAddressQuery query) throws IOException {
+		return workAddressService.queryForList(query); 
 	}
 	
 	@RequestMapping(mapping = { "/all", "/m/all" })
@@ -51,13 +58,19 @@ public class WorkAddressController {
 	
 	@RequestMapping(mapping = { "/save_or_update", "/m/save_or_update" })
 	public WorkAddress saveOrUpdate(WorkAddress form) {
+		if (StringUtil.isBlank(form.getCode())) {
+			form.setCode(idgen.getUUID58().toString());
+		}
 		return workAddressService.saveOrUpdate(form);
 	}
 	
 	@RequestMapping(mapping = { "/delete", "/m/delete" })
 	public int delete(String ids) {
-		List<Long> list = StringUtil.split(Long.class, ids, ",");
-		return workAddressService.deleteById(list.toArray(new Long[list.size()]));
+		if (StringUtil.isNotBlank(ids)) {
+			List<Long> list = StringUtil.split(Long.class, ids, ",");
+			return workAddressService.deleteById(list.toArray(new Long[list.size()]));
+		}
+		return 0;
 	}
 	
 }
